+++ app/src/main/java/fnsb/macstransit/RouteMatch/MarkedObject.java
@@ -1,0 +1,38 @@
+package fnsb.macstransit.RouteMatch;
+
+import com.google.android.gms.maps.model.Marker;
+
+/**
+ * Created by Spud on 2019-11-20 for the project: MACS Transit.
+ * <p>
+ * For the license, view the file titled LICENSE at the root of the project
+ *
+ * @version 1.0
+ * @since Beta 8.
+ */
+public class MarkedObject {
+
+	/**
+	 * TODO Documentation
+	 */
+	private Marker marker;
+
+	/**
+	 * TODO Documentation
+	 *
+	 * @return
+	 */
+	public Marker getMarker() {
+		return this.marker;
+	}
+
+	/**
+	 * TODO Documentation
+	 *
+	 * @param marker
+	 */
+	public void setMarker(Marker marker) {
+		this.marker = marker;
+	}
+
+}
